Fix/tm 1375 fixes dashboard learn more (#649)
* [TM-1375] fixes table dashboard profle

* [TM-1375] TM-1375 fix tooltips, routes, and add button

* [TM-1375] fixes linked in text TerraMatch’s monitoring, reporting, and verification system.

* [TM-1375] TM-1375 fix tooltips

* [TM-1375] fixes import

* [TM-1375] fixes import

---------
