+++ src/components/extensive/BlurContainer/BlurContainer.tsx
@@ -17,8 +17,8 @@ const BlurContainer = ({ isBlur, textInformation, children, ...props }: BlurCont
   return (
     <div className="relative w-full text-black">
       <div
-        className={`absolute left-0 top-0 flex h-full w-full items-center justify-center rounded-xl   ${
-          isBlur ? "z-10 bg-[#9d9a9a29] backdrop-blur-sm" : ""
+        className={`absolute left-0 top-0 flex h-full w-full items-center justify-center rounded-xl ${
+          isBlur ? "z-[1] bg-[#9d9a9a29] backdrop-blur-sm" : ""
         }`}
       >
         <When condition={isBlur && textInformation}>
