+++ src/components/extensive/PageElements/Card/PageCard.tsx
@@ -77,6 +77,7 @@ const PageCard = ({
                 <When condition={!!tooltip}>
                   &nbsp;
                   <ToolTip
+                    className="normal-case"
                     content={tooltip || ""}
                     width={widthTooltip || "w-44 lg:w-52"}
                     title={title}
